+++ libraries/common/src/main/java/androidx/media3/common/util/CodecSpecificDataUtil.java
@@ -38,6 +38,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+import java.util.Objects;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -1320,6 +1321,36 @@ public final class CodecSpecificDataUtil {
       return null;
     }
     return new MediaCodecProfileAndLevel(profile, level);
+  }
+
+  /**
+   * Returns a Dolby Vision base layer codec MIME type of the provided {@link Format}.
+   *
+   * @param format The media format.
+   * @return A Dolby Vision base layer MIME type, or {@code null} if a Dolby Vision profile is not
+   *     identified.
+   */
+  @Nullable
+  public static String getDolbyVisionBaseLayerMimeType(Format format) {
+    if (!Objects.equals(format.sampleMimeType, MimeTypes.VIDEO_DOLBY_VISION)) {
+      return null;
+    }
+    @Nullable Pair<Integer, Integer> codecProfileAndLevel = getCodecProfileAndLevel(format);
+    if (codecProfileAndLevel == null) {
+      return null;
+    }
+    switch (codecProfileAndLevel.first) {
+      case CodecProfileLevel.DolbyVisionProfileDvheDtr: // profile 4
+      case CodecProfileLevel.DolbyVisionProfileDvheStn: // profile 5
+      case CodecProfileLevel.DolbyVisionProfileDvheSt: // profile 8
+        return MimeTypes.VIDEO_H265;
+      case CodecProfileLevel.DolbyVisionProfileDvavSe: // profile 9
+        return MimeTypes.VIDEO_H264;
+      case CodecProfileLevel.DolbyVisionProfileDvav110: // profile 10
+        return MimeTypes.VIDEO_AV1;
+      default:
+        return null;
+    }
   }
 
   /** Returns H263 profile and level from codec string. */